Add a dedicated /proposals route
