+++ src/app/proposals/page.tsx
@@ -1,0 +1,6 @@
+import ProposalsHome from "@/components/Proposals/ProposalsHome";
+
+export const revalidate = 60;
+
+export { generateMetadata } from "../page";
+export default ProposalsHome;
